pyproject.toml
[project]
name = "friday-tony-stark-demo"
version = "0.1.0"
description = "Friday — A Tony Stark-inspired MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "httpx",
    "livekit-agents[deepgram,groq,openai,sarvam,silero]>=1.5.1",
    "livekit-plugins-google>=1.5.1",
    "python-dotenv",
    "openwakeword",
    "pyaudio",
    "livekit-api",
    "pytest",
    "ddgs",          # web search backend used by search_web tool
    "keyboard",      # global hotkeys (kill switch, restart)
    "resemblyzer",   # speaker verification
]

[project.scripts]
friday = "server:main"
friday_voice = "agent_friday:dev"
friday_start = "friday_launcher:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server.py"]
